refactor: split CLI into command modules
Split app.py (487 lines) into focused command modules:

- app.py (114 lines) — root app, global callback, init, sub-app registration
- eval_commands.py (108 lines) — eval run, eval compare
- report_commands.py (66 lines) — report, report show/export stubs
- browse_commands.py (152 lines) — bundle list/show, run list/show/view
- config_commands.py (56 lines) — config cicd, config validate/show stubs
- planned.py (57 lines) — dataset, monitor, trace, model, agent stubs
- _planned.py (12 lines) — shared planned command helper

No behavior changes. All 96 tests pass.
